# backend/docs/openapi.yaml
openapi: 3.0.0
info:
  title: Todo API
  description: |
    REST API for Todo App with gamification features.
    
    ## Features
    - User authentication (JWT-based)
    - Todo CRUD operations with optimistic locking
    - User connections for shared todos
    - Gamification system (XP, levels, badges)
    - Real-time synchronization
    
    ## Performance Targets
    - API Response Time: < 200ms (p95)
    - Authentication: < 100ms (p95)
  version: 1.0.0
  contact:
    name: API Support
    email: api@todoapp.com
servers:
  - url: http://localhost:8080
    description: Local development server
  - url: https://api.todoapp.com
    description: Production server
...
paths:
...
  /api/v1/connections:
    get:
      tags:
        - Connections
      summary: List user connections
      operationId: listConnections
      security:
        - BearerAuth: []
      responses:
        '200':
          description: List of connections
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Connection'
  /api/v1/connections/invite:
    post:
      tags:
        - Connections
      summary: Create connection invitation
      operationId: createInvitation
      security:
        - BearerAuth: []
      responses:
        '201':
          description: Invitation created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/InvitationResponse'
  /api/v1/connections/invite/{token}:
    parameters:
      - name: token
        in: path
        required: true
        schema:
          type: string

    get:
      tags:
        - Connections
      summary: Validate invitation token
      operationId: validateInvitation
      responses:
        '200':
          description: Invitation is valid

    post:
      tags:
        - Connections
      summary: Accept invitation
      operationId: acceptInvitation
      security:
        - BearerAuth: []
      responses:
        '204':
          description: Invitation accepted
  /api/v1/users/me/stats:
    get:
      tags:
        - Gamification
      summary: Get user gamification stats
      operationId: getUserStats
      security:
        - BearerAuth: []
      responses:
        '200':
          description: User stats retrieved
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserStats'
  /api/v1/users/me/history:
    get:
      tags:
        - Gamification
      summary: Get points history
      operationId: getPointsHistory
      security:
        - BearerAuth: []
      parameters:
        - name: limit
          in: query
          schema:
            type: integer
            default: 50
            maximum: 100
      responses:
        '200':
          description: Points history retrieved
          content:
            application/json:
              schema:
                type: object
                properties:
                  history:
                    type: array
                    items:
                      $ref: '#/components/schemas/PointsTransaction'
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    Connection:
      type: object
      properties:
        id:
          type: string
          format: uuid
        user_a_id:
          type: string
          format: uuid
        user_b_id:
          type: string
          format: uuid
        status:
          type: string
          enum: [pending, accepted, rejected]
        created_at:
          type: string
          format: date-time
    InvitationResponse:
      type: object
      properties:
        connection_id:
          type: string
          format: uuid
        token:
          type: string
        invitation_link:
          type: string
        expires_at:
          type: string
          format: date-time
    UserStats:
      type: object
      properties:
        points:
          type: integer
        level:
          type: integer
        level_name:
          type: string
        streak:
          type: integer
        total_todos_completed:
          type: integer
        progress_to_next_level:
          type: number
        xp_for_next_level:
          type: integer
    PointsTransaction:
      type: object
      properties:
        id:
          type: string
          format: uuid
        points:
          type: integer
        reason:
          type: string
        created_at:
          type: string
          format: date-time
...